feat: useTokenDisabled hook
