+++ packages/dapp/src/components/warnings/RebasingTokenWarning.jsx
@@ -55,18 +55,19 @@ const ExceptionsLink = ({ msg }) => (
   </Link>
 );
 
-export const RebasingTokenWarning = ({ token, noShadow = false }) => (
-  <Flex align="center" direction="column" w="100%" mb="4">
-    <Alert
-      status="error"
-      borderRadius={5}
-      boxShadow={noShadow ? 'none' : '0px 1rem 2rem rgba(204, 218, 238, 0.8)'}
-    >
-      <AlertIcon minWidth="20px" />
-      <Text fontSize="small">
-        {token.symbol} is a rebasing token and cannot be bridged.{' '}
-        <ExceptionsLink msg="Learn more" />.
-      </Text>
-    </Alert>
-  </Flex>
-);
+export const RebasingTokenWarning = ({ token, noShadow = false }) =>
+  isRebasingToken(token) ? (
+    <Flex align="center" direction="column" w="100%" mb="4">
+      <Alert
+        status="error"
+        borderRadius={5}
+        boxShadow={noShadow ? 'none' : '0px 1rem 2rem rgba(204, 218, 238, 0.8)'}
+      >
+        <AlertIcon minWidth="20px" />
+        <Text fontSize="small">
+          {token.symbol} is a rebasing token and cannot be bridged.{' '}
+          <ExceptionsLink msg="Learn more" />.
+        </Text>
+      </Alert>
+    </Flex>
+  ) : null;
